pyproject.toml
[tool.poetry]
name = "cardetection"
version = "0.1.0"
description = ""
authors = ["anatoly <t.vlasov23@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.9.2"
clearml = "^1.16.5"
pillow = "^11.0.0"
mlflow = "^2.17.2"
ultralytics = "^8.3.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

